Sharing of Directions, will now properly save the re-geocoded point location and not the original location.
Use case: Get directions to a place such as a trail or park, in which the directions are to someplace other than the point (e.g. parking lot). Then share the route. The re-geocoded location (best parking lot) was not being reused, but rather the actual location (e.g. centroid of trail).